requirements.txt
python-telegram-bot==13.7
plexapi==4.13.2
python-dotenv==0.19.2
psutil==5.9.0
requests==2.26.0
